+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Use `texlab.diagnosticsDelay` to debounce diagnostics updates with the goal of reducing load of the client
+  ([#1551](https://github.com/latex-lsp/texlab/issues/1551))
+
 ## [5.26.0] - 2026-06-25
 
 ### Added